% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_derivative.R
\name{time_derivative}
\alias{time_derivative}
\title{Title}
\usage{
time_derivative(
  data,
  column = "",
  order_by = NULL,
  time_step = 1,
  logarithmic = FALSE
)
}
\arguments{
\item{data}{}

\item{column}{}

\item{order_by}{}

\item{time_step}{}

\item{logarithmic}{}
}
\description{
Title
}
